Based on the JSON lecture, it seems I am missing some steps for JSON high score implementation, but I don't yet understand clearly what I'm doing wrong :(
Conceptually I think it makes sense to write the name and score data to a JSON file and then use a key to iterate through and check each value and if the current score is  higher than any of the high scores then write the entire node, including the name and it's corresponding value to the high scores list.

However, the syntax we were given in class keeps throwing errors. Documentation, forums and youtube haven't helped much since everyone does things differently and it feels like I'm wasting time now. I'll try and ask for help after we get back from Spring Break.